Removed final keyword from Model constructor
When working with unit test frameworks like PHPSpec and PHPUnit, it's impossible to mock/stub entities that inherit from Phalcon\Mvc\Model, since the framework is not able to use it's own constructor into the extended class. When mocking or stubbing, you'll get an error like this:

```
Fatal error: Cannot override final method Phalcon\Mvc\Model::__construct()
```

This is a serious problem for any large project with a serious approach to code quality. I'm sure this causing a lot of problems for other developers and it's a very simple change. There's no reason for the Model constructor to be final.