Allow cached keys to be disabled
Cached keys to allow flush is a big performance killer when a lot of keys are created. A timestamp-based management should be implemented (or just an increment), using only one small key.
For now, this change allows to disable the cached keys, enabling the developer to manage this on his own.