Extract handler class name construction from Dispatcher::dispatch
The method `Phalcon\Mvc\Dispatcher::getControllerClass()` did not work
as it referenced a method at runtime that doesn't exist. This commit
fixes that by refactoring the building of the handler class name from
Dispatcher::dispatch into a getControllerClass method.

The method getControllerClass depends on the resolving of defaults to
work correctly which isn't great, however fixing this will require
reworking how defaults are set (which is outside the scope of this
change). As-is this won't change the current behaviour of
`Dispatcher::dispatch`, but some incorrect behaviour might be possible
if calling getControllerClass outside of the dispatch loop.