ORM: Case-insensitive column map option
Added a global ORM setting that allows to try to find a key in column
map case-insensitively if it was not found by an exact match at first
time. Can be a real fix for many Oracle bugs, because Phalcon does not
expect that column names are converted to uppercase by Oracle in the
result.