feat(Micro): Response handler
What:
Adds the possibility to create a response handler, that will handle the
response instead of the default hard coded micro handler.

Why:
Before the reponse format had to be coded into every route.
Now is it possible to create a custom rosponse handler, that will format
the data in custom created function.