Update memory.zep
In the documentation for addRole it says accessInherits can be a string or array. But that functionality is not implemented in addInherit.

I added to bug fix branch as it doesn't break BC - even if it does allow array to be used (when they could) -- I assumed this would be okay since the docblock indicated arrays are ok. If not, I will redo on new branch.

Related to Issue #11361