Update Injectable.zep
When no container is set in the Injectable class, it always call the function Di::getDefault(). When you call the __get() a lot, this is just a repeated call. So my proposal is to set the container property to the default DI when not DI is set. 

It's a little bit extra performance.