Update Form.zep
The uniqueness validator fails when using an entity in a form. In this form I assign a validator object. One of the rules in the validator is that the "name" field must be unique.

Works great but when I use this form to edit, it has a persisted entity. So the uniqueness validator checks that their is no record of that name and doesn't belong to the primary key field. 

Only problem is that the validator doesn't receive the entity object that is assigned to the form. It receives the entity that is passed to the method "isValid($data, $entity)". 

This one line added to the isValid function, passes the form entity as the entity to validation.